requirements.txt
Flask==3.1.0
flask-cors==6.0.1
psycopg2-binary==2.9.11
python-dotenv==1.0.1
firebase-admin==6.4.0
anthropic==0.39.0
httpx==0.27.2
httpcore==1.0.6